Avoid removing preinstalled snaps on core
This change prevents to remove preinstalled snaps. This is usefull to
test gadgets such as the caracalla.

i.e. when the modem-manager snap is removed, it is causing the
connection to the device is broken in the caracalla.

+++ tests/lib/reset.sh
@@ -97,7 +97,9 @@ reset_all_snap() {
             "bin" | "$gadget_name" | "$kernel_name" | core | README)
                 ;;
             *)
-                snap remove "$snap"
+                if ! echo "$SKIP_REMOVE_SNAPS" | grep -q -w $snap; then
+                    snap remove "$snap"
+                fi
                 ;;
         esac
     done
